Warehouse
In a typical warehouse application, users should be able to add items to the inventory. Deleting items from the inventory should be restricted to users with the necessary permissions. Additionally, users should have the capability to retrieve items based on specific filters. They should also be able to modify the details of existing items. Finally, the application should allow users to view the entire inventory.